useBounds hook
