+++ components/map-context/map.js
@@ -1,41 +1,12 @@
-import React, {useState, useCallback, useEffect, useRef} from 'react'
+import React, {useRef} from 'react'
 import PropTypes from 'prop-types'
-import ReactMapGL, {WebMercatorViewport} from 'react-map-gl'
+import ReactMapGL from 'react-map-gl'
 
-import geo from '../../geo.json'
-
-const defaultViewport = {
-  latitude: 46.9,
-  longitude: 1.7,
-  zoom: 5
-}
+import useBounds from '../hooks/bounds'
 
 const Map = ({code, interactiveLayerIds, hideAttribution, onHover, onClick, children}) => {
   const mapRef = useRef()
-
-  const [viewport, setViewport] = useState(defaultViewport)
-
-  const handleResize = useCallback(() => {
-    if (mapRef && mapRef.current) {
-      const {width, height} = mapRef.current.getBoundingClientRect()
-
-      const {bbox} = geo[code]
-      const padding = width > 50 && height > 50 ? 20 : 0
-      const viewport = new WebMercatorViewport({width, height})
-        .fitBounds([[bbox[0], bbox[1]], [bbox[2], bbox[3]]], {padding})
-
-      setViewport(viewport)
-    }
-  }, [mapRef, code])
-
-  useEffect(() => {
-    handleResize()
-    window.addEventListener('resize', handleResize)
-
-    return () => {
-      window.removeEventListener('resize', handleResize)
-    }
-  }, [handleResize])
+  const viewport = useBounds(mapRef, code)
 
   return (
     <div ref={mapRef} className='react-map-container'>
